Significant changes related to 16-way design, which is now partially tested
-Extra checking in detailedPedigree validation code
-Fixed bugs in mathematica code, including one array being the wrong
dimensions
-Changes throughout to allow 16 parents
-Fixed warning in estimateRFCheckFunnels.cpp
-Changed encoded funnel type to 8 bytes, and use 4 bits for each entry,
rather than 3
-Fixed bug in getFunnel
-Fixed bugs in the 16-way probability calculations
-Shifted a huge array to its own file

+++ src/impute.cpp
@@ -10,7 +10,7 @@
 template<bool hasLOD, bool hasLKHD> bool imputeInternal(unsigned char* theta, std::vector<double>& levels, double* lod, double* lkhd, std::vector<int>& markersThisGroup, std::string& error, std::function<void(unsigned long, unsigned long)> statusFunction)
 {
 	unsigned long done = 0;
-	unsigned long total = markersThisGroup.size();
+	unsigned long total = (unsigned long)markersThisGroup.size();
 	unsigned long doneThreadZero = 0;
 	bool hasError = false;
 	//This is a marker row
